+++ Lesson10/src/main/java/by/tms/Task1.java
@@ -1,15 +1,15 @@
 package by.tms;
 
 import by.tms.model.Robot;
-import by.tms.model.samsung.SamsungHand;
-import by.tms.model.samsung.SamsungHead;
-import by.tms.model.samsung.SamsungLeg;
-import by.tms.model.sony.SonyHand;
-import by.tms.model.sony.SonyHead;
-import by.tms.model.sony.SonyLeg;
-import by.tms.model.toshiba.ToshibaHand;
-import by.tms.model.toshiba.ToshibaHead;
-import by.tms.model.toshiba.ToshibaLeg;
+import by.tms.model.robotparts.samsung.SamsungHand;
+import by.tms.model.robotparts.samsung.SamsungHead;
+import by.tms.model.robotparts.samsung.SamsungLeg;
+import by.tms.model.robotparts.sony.SonyHand;
+import by.tms.model.robotparts.sony.SonyHead;
+import by.tms.model.robotparts.sony.SonyLeg;
+import by.tms.model.robotparts.toshiba.ToshibaHand;
+import by.tms.model.robotparts.toshiba.ToshibaHead;
+import by.tms.model.robotparts.toshiba.ToshibaLeg;
 
 import java.util.ArrayList;
