+++ Lesson10/src/main/java/by/tms/model/robotparts/samsung/SamsungHead.java
@@ -1,4 +1,4 @@
-package by.tms.model.samsung;
+package by.tms.model.robotparts.samsung;
 
 import by.tms.interfaces.IHead;
 import lombok.Getter;
