+++ Lesson10/src/main/java/by/tms/model/robotparts/sony/SonyHand.java
@@ -1,4 +1,4 @@
-package by.tms.model.sony;
+package by.tms.model.robotparts.sony;
 
 import by.tms.interfaces.IHand;
 import lombok.Getter;
