+++ Lesson10/src/main/java/by/tms/model/robotparts/toshiba/ToshibaHand.java
@@ -1,4 +1,4 @@
-package by.tms.model.toshiba;
+package by.tms.model.robotparts.toshiba;
 
 import by.tms.interfaces.IHand;
 import lombok.Getter;
