+++ Lesson10/src/main/java/by/tms/model/robotparts/toshiba/ToshibaHead.java
@@ -1,4 +1,4 @@
-package by.tms.model.toshiba;
+package by.tms.model.robotparts.toshiba;
 
 import by.tms.interfaces.IHead;
 import lombok.Getter;
