+++ Lesson10/src/main/java/by/tms/model/robotparts/toshiba/ToshibaLeg.java
@@ -1,4 +1,4 @@
-package by.tms.model.sony;
+package by.tms.model.robotparts.toshiba;
 
 import by.tms.interfaces.ILeg;
 import lombok.Getter;
@@ -6,11 +6,11 @@ import lombok.ToString;
 
 @Getter
 @ToString(onlyExplicitlyIncluded = true)
-public class SonyLeg implements ILeg {
-    private final int cost = 22;
+public class ToshibaLeg implements ILeg {
+    private final int cost = 16;
 
     @Override
     public String go() {
-        return "the leg moves cinematically";
+        return "the leg uses a wheel and a gyroscope";
     }
 }
